+++ mesh_protocol_dwm1001-dev/src/Neighborhood.c
@@ -38,7 +38,7 @@
 
 #include "../include/Neighborhood.h"
 
-static bool removeNeighbor(Node node, int8_t neighborId);
+static bool removeNeighbor(Node node, uint16_t neighborId);
 
 Neighborhood Neighborhood_Create() {
   Neighborhood self = calloc(1, sizeof(NeighborhoodStruct));
@@ -56,11 +56,11 @@ Neighborhood Neighborhood_Create() {
   return self;
 };
 
-void Neighborhood_AddOrUpdateOneHopNeighbor(Node node, int8_t id) {
+void Neighborhood_AddOrUpdateOneHopNeighbor(Node node, uint16_t id) {
   // find neighbor in array (if present)
   int64_t localTime = ProtocolClock_GetLocalTime(node->clock);
   int8_t currentNumNeighbors = node->neighborhood->numOneHopNeighbors;
-  int8_t idx = Util_Int8tArrayFindElement(&node->neighborhood->oneHopNeighbors[0], id, currentNumNeighbors);
+  uint16_t idx = Util_Int8tArrayFindElement(&node->neighborhood->oneHopNeighbors[0], id, currentNumNeighbors);
 
   if (idx == -1) {
     // neighbor not in array yet, so add neighbor
@@ -75,7 +75,7 @@ void Neighborhood_AddOrUpdateOneHopNeighbor(Node node, int8_t id) {
   };
 };
 
-int8_t Neighborhood_GetOneHopNeighbors(Node node, int8_t *buffer, int8_t size) {
+int8_t Neighborhood_GetOneHopNeighbors(Node node, uint16_t *buffer, int8_t size) {
   if(size < node->neighborhood->numOneHopNeighbors) {
     // size of buffer is too small
     return -1;
@@ -89,7 +89,7 @@ int8_t Neighborhood_GetOneHopNeighbors(Node node, int8_t *buffer, int8_t size) {
 
 void Neighborhood_RemoveAbsentNeighbors(Node node) {
  
-  int8_t absentNeighbors[MAX_NUM_NODES - 1];
+  uint16_t absentNeighbors[MAX_NUM_NODES - 1];
   int8_t numAbsentNeighbors = 0;
   int64_t localTime = ProtocolClock_GetLocalTime(node->clock);
   
@@ -108,18 +108,18 @@ void Neighborhood_RemoveAbsentNeighbors(Node node) {
 
 };
 
-void Neighborhood_UpdateRanging(Node node, int8_t id, int64_t updateTime, double distance) {
+void Neighborhood_UpdateRanging(Node node, uint16_t id, int64_t updateTime, double distance) {
   // update the time when the last time ranging was done with this particular neighbor
 
   // find the index of the neighbor in the array
-  int16_t idx = Util_Int8tArrayFindElement(&node->neighborhood->oneHopNeighbors[0], id, node->neighborhood->numOneHopNeighbors);
+  uint16_t idx = Util_Int8tArrayFindElement(&node->neighborhood->oneHopNeighbors[0], id, node->neighborhood->numOneHopNeighbors);
   // update time
   node->neighborhood->oneHopNeighborsLastRanging[idx] = updateTime;
   // update distance
   node->neighborhood->oneHopNeighborsLastDistance[idx] = distance;
 };
 
-int8_t Neighborhood_GetNextRangingNeighbor(Node node) {
+uint16_t Neighborhood_GetNextRangingNeighbor(Node node) {
   // find the neighbor that was not ranged for the longest time
   if (node->neighborhood->numOneHopNeighbors == 0) {
     return -1;
@@ -137,7 +137,7 @@ int8_t Neighborhood_GetNextRangingNeighbor(Node node) {
   return node->neighborhood->oneHopNeighbors[minIdx];
 };
 
-int8_t Neighborhood_GetNewestNeighbor(Node node) {
+uint16_t Neighborhood_GetNewestNeighbor(Node node) {
   // find index of the neighbor that was added last
   int16_t idxNewestNeighbor = Util_Int64tFindIdxOfMaximumInArray(&node->neighborhood->oneHopNeighborsJoinedTime[0], node->neighborhood->numOneHopNeighbors);
 
@@ -161,9 +161,9 @@ int64_t Neighborhood_GetTimeWhenNewestNeighborJoined(Node node) {
   return node->neighborhood->oneHopNeighborsJoinedTime[idxNewestNeighbor];
 };
 
-static bool removeNeighbor(Node node, int8_t neighborId) {
+static bool removeNeighbor(Node node, uint16_t neighborId) {
   // find index of neighbor that should be removed
-  int8_t idx = Util_Int8tArrayFindElement(&node->neighborhood->oneHopNeighbors[0], neighborId, node->neighborhood->numOneHopNeighbors);
+  uint16_t idx = Util_Int8tArrayFindElement(&node->neighborhood->oneHopNeighbors[0], neighborId, node->neighborhood->numOneHopNeighbors);
   if (idx == -1) {
     // id not found
     return false;
